Move functions to tools.
